+++ FadeBasic/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.59] - 2026-04-1
+### Fixed
+- debugger shows hover values when looking at struct fields and array fields
+- debugger statement eval shows output
+
 ## [0.0.58] - 2026-03-24
 ### Fixed
 - debugger handles suspended vm correctly